add option to include CRs and VRs in the full likelihood
Option introduced in the `.info` file of the analysis e.g.
```
  <pyhf id="RegionC">
    <name>atlas_susy_2018_31_SRC.json</name>
    <regions>
      <channel name="CRtt_cuts"> </channel>
      <channel name="SR_metsigST">
      SRC_22
      SRC_24
      SRC_26
      SRC_28
      </channel>
      <channel name="VRzAlt_cuts" is_included="True"> </channel>
      <channel name="CRz_cuts" is_included="True"> </channel>
      <channel name="VRtt_cuts"> </channel>
      <channel name="VRz_cuts"> </channel>
    </regions>
  </pyhf>
```
If there is no data present in the channel code will look for `is_included`
tag which can be "True", "yes", "1". If true the channel will be included
in the likelihood calculation if not it will be removed. If `is_included`
is not present and channel does not have any data, it will be removed by
default.